+++ package.json
@@ -1,15 +1,20 @@
 {
-  "name": "tweedentity-api",
+  "name": "tweeapi",
+  "description": "A simple API to retrieve and verify a signed tweet for Tweedentity",
   "license": "MIT",
-  "version": "0.0.1",
+  "version": "1.0.0",
+  "main": "handler.js",
+  "author": "Francesco Sullo <francesco@sullo.co>",
   "scripts": {
-    "start": "node ."
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "devDependencies": {
+    "serverless-domain-manager": "^2.3.5",
+    "serverless-offline": "^3.18.0"
   },
   "dependencies": {
     "cheerio": "^1.0.0-rc.2",
     "ethereumjs-util": "^5.1.3",
-    "superagent": "^3.8.2",
-    "web3": "^1.0.0-beta.29"
-  },
-  "devDependencies": {}
+    "superagent": "^3.8.2"
+  }
 }
